[WEBSITE-219] - Properly navigate to dependencies
The plugin being displayed is in the redux state. This is loaded when
the component is mounted. The plugin to load is read via the router
parameter `pluginName`. But nothing was monitoring subsequent changes
to said parameter. The solution is to implement
`componentWillReceiveProps` and determine if the parameter has indeed
changed, and if so get the new plugin. This elegantly handles forward
and back navigation.